feat(general): A bit of refactoring
