+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./Card";
+export * from "./CopyToClipboardBtn";
+export * from "./Layout";
+export * from "./PoweredByVercel";
+export * from "./SearchBar";
+export * from "./ThemeSwitcher";
